Fix: page 디렉토리 명칭 통일
