+++ src/App.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import { Provider } from 'react-redux';
 import store from './redux/store';
-import HomePage from './pages/HomePage';
+import HomePage from './pages/Home';
 import CreateNicknamePage from './pages/CreateNickname';
 import CreateHabitPage from './pages/CreateHabit';
 import MyHabitPage from './pages/MyHabit';
